[region-isolation] Reformat transfer non sendable so that utility code is next to the routines that use the utility.
NFC.

Before this commit, the diagnostic emission was split into utilities that were
used by one large emitter class. This could get confusing since there was some
duplication of names in between the utilities. Now instead, each part of the
main diagnostic emitter is in the MARK: section of the utilities that it uses.

Just making the pass easier for me to reason about.