Sema: Try a little harder to infer associated types to generic parameters if all else fails
If we have an abstract witness, we don't attempt a generic parameter
binding at all. But if simplifying the abstract witness failed, we
should still attempt it.

This would be cleaner as a disjunction in the solver but I want
to change behavior as little as possible, so this adds a new fallback
that we run when all else fails.

Fixes rdar://problem/123345520.