[Distributed] Undo new record and mangling scheme for dist.p.witnesses
This change undoes an approach we thought viable to encode distributed
actor remote targets, as encoding their protocol method names. This also
needed extra metadata sections and tricky "find if this is a witness or
not" methods;

Instead, we will be mangling the concrete names; and special handle
proxy types that are prefixed with `$`. This way we will not need extra
metadata records, and can handle generic calls more properly.

This XFAILs one test which was testing new functionality, that is
pending now until we implement the new mangling scheme.