[Concurrency] SerialExecutor.checkIsolated
Allow serial executors to perform a "last resort" check before crashing
in calls like preconditionIsolated and assumeIsolated.

This allows custom excecutors to use external information to claim that
while the synchronous code is not executing in a Swift Task, it IS
executing on the apropriate thread or queue the code is expecting.

add swift_task_checkIsolatedImpl to CooperativeGlobalExecutor