[DebugInfo] Emit mangled name of opaque container for debug info
With the new SpecifiedOf attribute, consumers of debug info will follow
the specification link and copy over any attribute that are not
specified in the current type. For this reason, emit the opaque
container's mangled name as well, as it may have a different name from
the type it specifies.