Sema: Fix checkTypeWitness() for Objective-C protocol compositions
I forgot that a protocol composition (C & P) can satisfy a superclass
requirement [T : D] in one narrow case implemented in
TypeBase::isExactSuperclassOf():

- D is a superclass of C
- P is an @objc protocol
- C is declared in Objective-C

This case was ruled out here because the code assumed the type witness
had to be a concrete class or archetype to satisfy a superclass
requirement.

Fixes rdar://problem/123543200.