[DebugInfo] Fix dynamic self debug loc
The instructions related to producing DynamicSelf metadata are always placed at
the start of the function. This creates an issue: which debug location should be
used in order to preserve a sane line table? The current implementation just
uses w/e debug location is present in the new insertion point, which could, for
example, be a debug_declare whose location is not suitable for these purposes.

This patch addresses the issue by removing the debug location of instructions
associated with producing DynamicSelfMetadata.

rdar://120408665