[Runtime] Fix KeyPath equality with captures of differing types.
Check that the captures have the same type by comparing their witnesses before calling their equality function.

Key paths can capture subscript arguments, as long as they're Hashasble. The implementation of AnyKeyPath's == function uses the captured witnesses to perform the comparison. However, the captured types can differ if the subscript is generic, and equality checking didn't accommodate for that. As a result, the second capture's memory would be interpreted as the first capture's type, resulting in false positives or crashes depending on exactly which types and data were involved.

rdar://122630313