test: Repair isolated_any_metadata.sil for Wasm CI
The test was failing on the Wasm CI because `%IRGenFileCheck` is
configured for the target platform (in this case wasm32-unknown-wasi),
but the test used it to check the output for other platforms. The
mismatch was not revealed on the other CI because they are all 64-bit,
but the Wasm CI is 32-bit.