test: Repair isolated_any_metadata.sil for Wasm CI (#72013)
The test was failing on the Wasm CI because `%IRGenFileCheck` is configured for the target platform (in this case wasm32-unknown-wasi), but the test used it to check the output for other platforms. The mismatch was not revealed on the other CI because they are all 64-bit, but the 32-bit Wasm CI revealed it. https://ci.swift.org/job/oss-swift-pr-test-crosscompile-wasm-ubuntu-20_04/66/

Adjust a test case added by https://github.com/apple/swift/pull/71970 to repair Wasm CI