[IRGen] Require same copyability to reuse layout.
Previously, in place of the layout of an aggregate which consisted of
exactly one field, the layout of that one field would be used.  This is
incorrect in the case that the aggregate is noncopyable but the field is
not.  Here, it's checked that the move-only-ness of the field matches
the move-only-ness of the aggregate.

rdar://123197751