[concurrency] Eliminate a double error emission caused by erroring on an autoclosure and the callexpr it calls.
Specifically, when one is emitting a non-Sendable diagnostic for an async let,
we will emit an error both for the async let and for the call expr inside the
async let. This just puts in a check so that we do not emit the error for the
autoclosure if we can show that the autoclosure's child subexpr would have the
same type implying we would also emit the same error.